Working on secure & reliable message passing. WIP
This commit marks the beginning of a more secure and reliable message passing
architecture based on a single central object that manages sending messages for
a connection: The MessageSendQueue. This takes over the send queue management
duties from TCPConnection and handles queue and ACK management itself. It
survives a dying TCPConnection and fixes the issue of unavailable
send_message_to functions.

Issues that remain: State machine is still not working correctly combined with
TCPConnection, more complex scenarios fail.